Exam Service APIs Admin and User Routes Setup done

+++ backend/exam-service/controllers/exam.controller.js
@@ -3,7 +3,22 @@ import mongoose from "mongoose";
 // Get all Exams
 const getAllExams = async (req, res) => {
   try {
-    const exams = await Exam.find().populate("questionIds", "-correctOption");
+    const exams = await Exam.find().populate(
+      "questionIds",
+      "-correctOption -examId"
+    );
+    res.status(200).json(exams);
+  } catch (error) {
+    res
+      .status(404)
+      .json({ message: "Error fetching exams: ", error: error.message });
+  }
+};
+
+const getAllExamsWithAnswers = async (req, res) => {
+  try {
+    // Populate the questionIds field with all fields including correctOption and examId
+    const exams = await Exam.find().populate("questionIds", "-examId");
     res.status(200).json(exams);
   } catch (error) {
     res
@@ -12,13 +27,16 @@ const getAllExams = async (req, res) => {
   }
 };
 const getExamById = async (req, res) => {
-  const { id } = req.params;
   try {
+    const { id } = req.params;
     if (mongoose.Types.ObjectId.isValid(id) === false) {
       return res.status(400).json({ message: "Invalid Exam ID format" });
     }
 
-    const exam = await Exam.findById(id).populate("questionIds");
+    const exam = await Exam.findById(id).populate(
+      "questionIds",
+      "-correctOption -examId"
+    );
     if (!exam) {
       return res.status(404).json({ message: "Exam not found" });
     }
@@ -30,13 +48,13 @@ const getExamById = async (req, res) => {
   }
 };
 
-const getExamByIdWithQuestions = async (req, res) => {
-  const { id } = req.params;
+const getExamByIdWithAnswers = async (req, res) => {
   try {
+    const { id } = req.params;
     if (mongoose.Types.ObjectId.isValid(id) === false) {
       return res.status(400).json({ message: "Invalid Exam ID format" });
     }
-    const exam = await Exam.findById(id).populate("questionIds");
+    const exam = await Exam.findById(id).populate("questionIds", "-examId");
     if (!exam) {
       return res.status(404).json({ message: "Exam not found" });
     }
@@ -49,40 +67,31 @@ const getExamByIdWithQuestions = async (req, res) => {
 };
 
 const createExam = async (req, res) => {
-  const { title, description, duration_minutes, start_time } = req.body;
-
-  if (!title || !description || !duration_minutes || !start_time) {
-    return res.status(400).json({ message: "All fields are required" });
-  }
-
-  const newExam = new Exam({
-    title,
-    description,
-    duration_minutes,
-    start_time: new Date(start_time), // Ensure start_time is a Date object
-  });
-
-  await newExam
-    .save()
-    .then((exam) => {
-      res.status(201).json(exam);
-    })
-    .catch((error) => {
-      res
-        .status(500)
-        .json({ message: "Error creating exam", error: error.message });
+  try {
+    const { title, description, duration_minutes, start_time } = req.body;
+    const newExam = new Exam({
+      title,
+      description,
+      duration_minutes,
+      start_time,
     });
+    await newExam.save();
+    res.status(201).json(newExam);
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Error creating exam", error: error.message });
+  }
 };
 const updateExam = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, duration_minutes } = req.body;
-
-  const exam = await Exam.findByIdAndUpdate(
-    id,
-    { title, description, duration_minutes },
-    { new: true }
-  );
   try {
+    const { id } = req.params;
+    const { title, description, duration_minutes, start_time } = req.body;
+    const exam = await Exam.findByIdAndUpdate(
+      id,
+      { title, description, duration_minutes, start_time },
+      { new: true }
+    );
     if (!exam) {
       return res.status(404).json({ message: "Exam not found" });
     }
@@ -94,10 +103,13 @@ const updateExam = async (req, res) => {
   }
 };
 const deleteExam = async (req, res) => {
-  const { id } = req.params;
-
-  const exam = await Exam.findByIdAndDelete(id);
   try {
+    const { id } = req.params;
+
+    if (mongoose.Types.ObjectId.isValid(id) === false) {
+      return res.status(400).json({ message: "Invalid Exam ID format" });
+    }
+    const exam = await Exam.findByIdAndDelete(id);
     if (!exam) {
       return res.status(404).json({ message: "Exam not found" });
     }
@@ -112,7 +124,8 @@ const deleteExam = async (req, res) => {
 export default {
   getAllExams,
   getExamById,
-  getExamByIdWithQuestions,
+  getAllExamsWithAnswers,
+  getExamByIdWithAnswers,
   createExam,
   updateExam,
   deleteExam,
